requirements.txt
python-dotenv==1.0.0
bs4==0.0.1
promptflow==1.1.1
promptflow-tools==1.0.0
openai==1.3.7
#openai==0.28.1
chainlit==0.7.501
keyrings.alt==5.0.0
langchain==0.0.350
langchain-community==0.0.7
langchain-core==0.1.4
faiss-cpu==1.7.4
pandas==2.1.4
matplotlib==3.8.2
geopandas==0.14.1
requests==2.31.0
pyautogen==0.2.0
#pyautogen==0.1.14
docker==7.0.0
pydotenv==0.0.7
streamlit==1.29.0
gunicorn==20.1.0 
pgvector==0.2.4
psycopg2_binary==2.9.9
pyogrio==0.7.2
hdx_python_api==6.2.1
pycountry==23.12.11
streamlit_code_editor==0.1.10
deepeval==0.20.97
